requirements.txt
chainlit
meilisearch
python-dotenv
langdetect
litellm
google-generativeai